[K/N] Pick minimal iOS/tvOS depending on the Xcode version
ld from Xcode 15 beta 5 incorrectly parses dylib version if it was built
with version < 12.0.

In the master branch we simply solved the problem by bumping
minOsVersion to 12.0. However, in 1.9.10 we have to support both Xcode 14
as well as Xcode 15. Thus, we a added a bit of logic that checks Xcode
version and sets minOsVersion depending on it.

^KT-60758