[GH] stop using buildSrc symlinks
This CL restructures playground projects not to use buildSrc to workaround a new Studio bug where it fails to
load projects with symlinks.

Instead of using a buildSrc symlink, we make the playground-plugin depend on AOSP's buildSrc, which in return
makes it available in the classpath when the playground-plugin is applied.

I also needed to change how playground-plugin is applied because we need to declare repositories to be able
to resolve the classpath and I didn't want to duplicate repositories across all settings files.

See: playground-common/configure-plugin-management.gradle for details

Bug: 295337712 ( https://issuetracker.google.com/issues/295337712 )
Idea Bug: https://youtrack.jetbrains.com/issue/IDEA-329756
Test: CI

This is an imported pull request from https://github.com/androidx/androidx/pull/603.

Resolves #603
Github-Pr-Head-Sha: e56629d09f906585efb622bfaefe277d107e46ac
GitOrigin-RevId: ebe535885618fa8eee52ccff3ef9ca2bdc1a29bd
Change-Id: I00d69ca18d1f6366c43ad6621a22b2e868907e07 ( https://android-review.git.corp.google.com/q/I00d69ca18d1f6366c43ad6621a22b2e868907e07 )