[Gradle] Fix duplicate rendering of FATAL-diagnostics
We throw FATAL-diagnostic right as it has been reported, but we also
store it. On the other hand, we schedule diagnostic launching regardless
of the project configuration result status - so in the coroutine, we'd
throw the FATAL-diagnostic for the second time.

This commit just filters our FATAL in that coroutine.

More "smart" solution would be to somehow track which diagnostics
are reported and which are not, but for now keep it simple