[Gradle] Make diagnostic reporting/handling logic more consistent (3/3)
Use two separate instances of the KotlinToolingDiagnosticCollector:
one for configuration phase, another for execution phase.

The reason why we have to go through such a weird hoop is because Gradle
properly serializes BuildService.Parameters only if they were configured
in registerIfAbsent-action. All consequent mutations to will not be
serialized for execution-phase. Therefore, after the configuration phase
we'd lose:
 a) all diagnostics (making correct deduplication impossible)
 b) isTransparent flag (making diagnostics reported from tasks to be
 swallowed)

So, instead we collect all the diagnostics during the configuration
phase in one instance of KotlinToolingDiagnosticCollector, and then
configure the second instance that pulls data from the first one.