Include projects with dependencies in Playground
Playground builds can declare a dependency on a project using
projectOrArtifact syntax intead of project, to depend on its
prebuilt rather than the actual project. This was a way to keep
github projects lean.

Over the years, this has been a major trigger for breaking builds
as it is easy for any engineer to add a dependency without thinking
about the github build.
Furthermore, with the gradle remote cache, we don't necessarily need
it that much (still do, but a lot less).

This CL significantly reduces the need for it by making github use
the project dependency resolution logic (PROJECT_PREFIX) from the
main build.

This makes the builds more similar to the main build but comes with
the disadvange of running too much stuff in CI, where we don't have
good machines. It also creates a problem where the tests of commonly
included projects run will multiple times for each build.
To address this problem, I've created new mechanism where we distinguish
between projects requested by the settings.gradle file vs projects
included because they are a dependency. To do so, PlaygroundExtension
passes down a parameter to the root project about the projects that
are requested. Subsequently, AndroidXPlaygroundRootImplPlugin sets up
the playgroundCIHostTests task to only depend on them.

I've replaced a bunch of projectOrArtfiact calls with project where
needed but didn't do a full cleanup. Unfortunately, we still cannot
remove all of them since some cannot be built on Github. But after a
cleanup (followup CL), using projectOrArtifact should be a one-off
case instead of the norm (and we can probably make the playground
extension do it instead of the project files).

Last but not least, this CL greatly simplifies settings files for
playground projects as they don't need to think about their
dependencies. Current syntax is not great for this new norm so I hope
to update the settings.gradle syntax in a followup CL (to make it more
explicit instead of iterating over all projects).

Bug: n/a
Test: https://github.com/androidx/androidx/actions/runs/6280777658

Change-Id: Ida6b649f7610a20f3626e38e444e5b21f4034cc6 ( https://android-review.git.corp.google.com/q/Ida6b649f7610a20f3626e38e444e5b21f4034cc6 )