Propagate Certain stability values in annotation
Modifies @StabilityInferred annotation to include information about known stability in (typeParameters + 1) bit. This fixes the issue with incremental compilation losing this information when part of the compilation unit is represented as stubs.

Also starts applying StabilityInferred annotation to internal classes, as they can be included from a different compilation module.

Test: Updated/added tests

Change-Id: Ib39a8b8d9a598a252ea7489686f264dab27b802a ( https://android-review.git.corp.google.com/q/Ib39a8b8d9a598a252ea7489686f264dab27b802a )