[FIR] Don't require override for `equals` in the real `Any`
But even with this fix the code still fails, because:
`IrLazyClass` has `declarations` that
generates IR for members, but for all
source classes we also call
`processClassMembers` which also generates
IR for members, so we get duplicates,
and later when we try to find some IR
function by condition and require there
to be a single instance, we crash.
We should probably first wait for Dimas
refactoring, maybe it fixes the issue.

^KT-62183
^KT-56076